Feature/#149 be api crud work shift management

+++ backend/internal/auth/repository/account.go
@@ -11,7 +11,7 @@ type AccountRepository interface {
 	GetAccountByEmail(ctx context.Context, email string) (hrmmodel.Account, error)
 	GetAccountByID(ctx context.Context, id string) (hrmmodel.Account, error)
 	UpdateAccount(ctx context.Context, account hrmmodel.Account) error
-	GetEmployeeByAccountID(ctx context.Context, accountID int) (hrmmodel.Employee, error)
+	GetEmployeeByAccountID(ctx context.Context, accountID int64) (hrmmodel.Employee, error)
 }
 
 // accountRepo struct chứa db connection và implement interface
@@ -45,7 +45,7 @@ func (r *accountRepo) GetAccountByID(ctx context.Context, id string) (hrmmodel.A
 func (r *accountRepo) UpdateAccount(ctx context.Context, account hrmmodel.Account) error {
 	return r.db.WithContext(ctx).Save(&account).Error
 }
-func (r *accountRepo) GetEmployeeByAccountID(ctx context.Context, accountID int) (hrmmodel.Employee, error) {
+func (r *accountRepo) GetEmployeeByAccountID(ctx context.Context, accountID int64) (hrmmodel.Employee, error) {
 	var employee hrmmodel.Employee
 	err := r.db.Raw(`SELECT
     e.employee_id,
